feat(fsweb): Convert proxy to Typescript
Also moves the specific middleware attachments into their own files so they can be applied separately, runs compile-web and the proxy in parallel in dev mode, and will serve the app regardless of whether the proxy is running in prod mode

To test with PirateShip prod, run:
yarn
yarn ship:init-web
yarn ship:build-web
yarn ship:server-prod

For dev:
yarn
yarn ship:init-web
yarn ship:server-dev

Right now there are no environments configured to work with the proxy, though, as "commercecloud" isn't a valid data source type in PirateShip.
