+++ packages/fsweb/package.json
@@ -9,8 +9,10 @@
     "build:size": "webpack --json | webpack-bundle-size-analyzer",
     "start": "webpack-dev-server -d --config ./webpack.config.js --mode development --inline --hot",
     "start-prod": "webpack-dev-server -d --config ./webpack.config.js --mode production",
-    "server-dev": "node ./proxy/server.dev.js",
-    "server-prod": "node ./proxy/server.prod.js"
+    "server-dev": "node ./dist/server.dev.js",
+    "server-prod": "node ./dist/server.prod.js",
+    "prepare": "tsc",
+    "tsc:watch": "tsc -w --preserveWatchOutput"
   },
   "dependencies": {
     "@babel/core": "^7.7.2",
@@ -50,6 +52,10 @@
     "webpack-hot-client": "^4.0.4",
     "webpack-manifest-plugin": "^2.0.2",
     "workbox-webpack-plugin": "^4.3.1"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.7",
+    "@types/express": "^4.17.8"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
