+++ packages/fsweb/src/lib/cors.ts
@@ -1,0 +1,8 @@
+import * as express from 'express';
+import * as cors from 'cors';
+
+export const addCors = (app: express.Express, apiConfig?: any): void => {
+  if (apiConfig?.corsOptions) {
+    app.use(cors(apiConfig.corsOptions));
+  }
+};
