+++ packages/pirateship/package.json
@@ -16,7 +16,8 @@
     "init:web": "node scripts/fixRNLink.js && flagship clean && flagship init web",
     "init:android-jetify": "jetify",
     "init:android-jetify-reverse": "jetify -r",
-    "start:server-dev": "cd web && yarn server-dev",
+    "start:dev-proxy": "cd web && yarn server-dev",
+    "start:server-dev": "run-p compile-web start:dev-proxy",
     "start:server-prod": "cd web && yarn server-prod",
     "prepare": "tsc",
     "start": "node node_modules/react-native/local-cli/cli.js start",
