+++ packages/pirateship/tsconfig.json
@@ -20,6 +20,7 @@
   },
   "exclude": [
     "**/*.test.ts",
-    "dist/"
+    "dist/",
+    "web/"
   ]
 }
